pyproject.toml
[tool.poetry]
name = "lck-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.6"
django = ">=5.1.4"
djangorestframework = ">=3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = ">=2.9.10"
djangorestframework-simplejwt = ">=5.3.1"
django-cors-headers = "^4.6.0"
pillow = "^11.1.0"
psycopg = "^3.2.4"
django-environ = ">=0.11.2"
isort = "^6.0.0"
postgres = ">=4.0"
uwsgi = ">=2.0.24,<2.1.0"
boto3 = ">=1.36.1"
django-soft-delete = "^1.0.18"
django-taggit = "^6.1.0"
django-taggit-templatetags2 = "^1.6.1"
mypy = "^1.14.1"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
mypy = "^1.14.1"
django-stubs = "^5.1.2"
coverage = "^7.6.10"
drf-spectacular = ">=0.28.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
python_version = "3.12"
strict = true
ignore_missing_imports = true
check_untyped_defs = true          # 타입이 없는 함수도 검사
show_error_codes = true            # 에러 코드 표시

[[tool.mypy.overrides]]
module = [
    "manage",
    "*.migrations.*",
    "config.*",
    "apps.users.models",
    "apps.teams.models",
    "apps.players.models",
    "apps.subscriptions.models"
]
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.dev"